add instance variable for geocoder address from lat
I had memoized, but not thinking it works like this. not totally
sure b/c i see that i'm actually only hitting it once right now,
for zip. i believe there are problems with consistently getting zip,
and there are probably more problems with the others for which there
are (dead) methods like street number. anyway now it is memoized
to the instance variable.