rework geocoding in meeting
we knew this was a problem. finally starting taking
a look and making some adjustments. first off,
a geocoder object (returned from searching with lat/lng)
has built in methods. the hash traversal was totally
unnecessary. yay rushed projects. :-) future todo, potentially,
and already added to waffle, is to look at which object
to use.. this is actually a search results list coming back.
ooh, maybe i save them as options. hmm. or be able to pull
them up later in admin console if i don't like what i see.

anyway, tests needed to be adjusted because they were stubbing
a method i got rid of. there is one method in meeting now to
get a geocoder object and store it as an instance variable.
another method will set address and zip based on lat/lng, if
lat/lng is available. at least theoretically. :-)

moving in the right direction people.