factor out meeting creation to meetings_creator
add delete logic here. not fully fleshed out,
but need to stop tonight and get this stuff
saved.

added deleted column to meeting
so now there is a visible, deleted, and approved.
will think about how to use these flags and how to rework
as needed.

extract.yml didn't get used yet. using fixtures. but it
was an interesting dump of current meetings.