remove arguments, make runner an instance method
can add arguments back in if there's a use case. otherwise,
you create a new meetings_creator with new raw)meetings,
execute the runner or the other methods as you see fit.
if you run the runner, you are left with ingredients for
a report in deleted, untouched, and created