modify to use deleted flag for Meeting
visible flag will be used by admin. when a meeting
is manually made visible, even if it is tagged as
deleted, it will remain visible. which is to say,
if it was tagged deleted and hidden, when the visible
flag is flipped, it will remain that way.

conversely, if a visible flag is flipped, the
meeting will go on its merry way being hidden,
even when update rake tasks are run.