fix sorting within distance groups
a select statement is used to get meetings within a range.
but this doesn't preserve order (assuming the meetings
are coming into this method in order, which may not
be the case).

at any rate, factored out the select process to make
it easier to see what was going on, and sorted
in the new method.