rework header/footer
Get rid of buttons on top. Get rid of meeting
count bubble. Change to new search on new search,
and search results on search index. Add little buttons
on the bottom to go back and for new searhc. Make
the buttons the same for both new search and on search
results, even though back and home have the same effect
from the new search screen.