ajaxify meeting searching
- remove global ajax:false setting from jqm init (not
sure this really did anything but it didn't hurt..
note that the other setting about link binding did
hurt when it was turned off.. it caused sytling to be
wiped when following links back to home and such)
- configure search form for data remote
- rework controller to give js response.. this enabled
cleanup so that no redirection to another action was
happening, which took away need for writing a search id
to the session which was done because of object hijacking
- had problems with the j rendered partial then missing some
of the jqm styling. this was most easily fixed by applying
the data theme to the single page we're using, which does
not change on ajax replace. TODO would be to see about this
page swapping business with jqm.. return a page, use some
function to change it out, do some animations, what have you.
- added javascript url-ing to change the url to -results-
- added a route so that refresh on this made up url would
take back to a new search. that necessitated bringing in
birds to bring in the spiders that were brought in to eat
the beatles.. no wait it just necessitated changing the url
of the new search as well so it wouldn't still say results
on a refresh in that manner