fix test errors by switching to factories
While working on meeting spec in preparation to scope open
and closed, broke a whole bunch of tests. The tests were
broken when run in a group, so it's my favorite issue
of unclean database.

Of course, db cleaner isn't running around each test or
test file, so there's that. But the reason they were working
before is due to the use of fixtures. Mixing factory girl
into the fix I'm thinking some things slipped through
somewhere..

So the fix is to use factory girl in every affected spec,
doing a manual clean of used tables before and after that
test file.