another wip commit on ajax menu updating
days menu is now telling the controller it has been changed,
then controller tells the rails controller, the rails controller
searches and creates a searches options which overrides as_json,
on page controller gives json back to days controller, it disables
menu items appropriately. whew.

now of course, we don't want the touch the menu that was the source
of the narrow. this was just to keep things slightly simpler. now
to abstract out the menu element, and get other menus involved.
then the source will filter those updates out