Gentle rake exit on no updates needed
Previously, if a no update determination was made based on
table update and remote update times, an error would be thrown
by MeetingsCreator because the raw_meetings it is passed would
be the bool false instead of a raw_meeting collection. Empty
collection deletes all the meetings.. Better to just bail
on false, for now.