Fix meetings with close parens
Still dealing with fallout from the bad strip command.. And this probably
won't finish it.. But it will kill the meetings that have a raw meeting
with an address that includes a non-whitespace character followed immediately
by an open paren, then it will kill their raw meetings, and since the
raw parser has been fixed, they will be created correctly.

But what else is wrong with the syncing? Not sure.