WIP - remove address elements from meeting
Specs are failing and they are uncovering usages of the address
elements on Meeting. So there's more work to be done here before
killing those elements.

Basically we're between two worlds... The addresses table is
in use with data copied from the address elements, but some
things are still accessing those elements.